feat(server): code generator version option on service settings
