+++ packages/amplication-server/src/core/serviceSettings/dto/CodeGeneratorVersionOptions.ts
@@ -1,0 +1,17 @@
+import { Field, ObjectType } from "@nestjs/graphql";
+import { CodeGeneratorVersionStrategy } from "./EnumCodeGeneratorVersionStrategy";
+
+@ObjectType({
+  isAbstract: true,
+})
+export class CodeGeneratorVersionOptions {
+  @Field(() => String, {
+    nullable: true,
+  })
+  version?: string;
+
+  @Field(() => CodeGeneratorVersionStrategy, {
+    nullable: true,
+  })
+  selectionStrategy?: CodeGeneratorVersionStrategy;
+}
