+++ packages/amplication-server/src/core/serviceSettings/dto/CodeGeneratorVersionOptionsInput.ts
@@ -1,0 +1,17 @@
+import { Field, InputType } from "@nestjs/graphql";
+import { CodeGeneratorVersionStrategy } from "./EnumCodeGeneratorVersionStrategy";
+
+@InputType({
+  isAbstract: true,
+})
+export class CodeGeneratorVersionOptionsInput {
+  @Field(() => String, {
+    nullable: true,
+  })
+  version?: string;
+
+  @Field(() => CodeGeneratorVersionStrategy, {
+    nullable: true,
+  })
+  selectionStrategy?: CodeGeneratorVersionStrategy;
+}
