+++ packages/amplication-server/src/core/serviceSettings/dto/EnumCodeGeneratorVersionStrategy.ts
@@ -1,0 +1,11 @@
+import { registerEnumType } from "@nestjs/graphql";
+
+export enum CodeGeneratorVersionStrategy {
+  LatestMajor = "LatestMajor",
+  LatestMinor = "LatestMinor",
+  Specific = "Specific",
+}
+
+registerEnumType(CodeGeneratorVersionStrategy, {
+  name: "CodeGeneratorVersionStrategy",
+});
