+++ packages/amplication-server/src/core/serviceSettings/dto/ServiceSettingsUpdateInput.ts
@@ -3,6 +3,7 @@ import { BlockUpdateInput } from "../../block/dto/BlockUpdateInput";
 import { AdminUISettingsUpdateInput } from "./AdminUISettingsUpdateInput";
 import { EnumAuthProviderType } from "./EnumAuthenticationProviderType";
 import { ServerSettingsUpdateInput } from "./ServerSettingsUpdateInput";
+import { CodeGeneratorVersionOptionsInput } from "./CodeGeneratorVersionOptionsInput";
 
 @InputType({
   isAbstract: true,
@@ -27,4 +28,9 @@ export class ServiceSettingsUpdateInput extends BlockUpdateInput {
     nullable: true,
   })
   authEntityName?: string;
+
+  @Field(() => CodeGeneratorVersionOptionsInput, {
+    nullable: false,
+  })
+  codeGeneratorVersionOptions!: CodeGeneratorVersionOptionsInput;
 }
